faq-chatbot-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
openai>=2.0.0
httpx==0.27.2
pinecone>=5.0.0
